Added logic to HLMS texture manager to prevent unhandled expecption when using compressed textures
The logic has been updated so if the HLMS Tex Mgr wants to convert an image to
a different format, but either one is a compressed texture then it falls back to
using the input image format along with a warning in the log.

Also updated the log levels for these warnings to normal from trivial.